+++ SharpC2.API/V1/Responses/DroneResponse.cs
@@ -11,9 +11,23 @@ namespace SharpC2.API.V1.Responses
         public string Username { get; set; }
         public string Process { get; set; }
         public int Pid { get; set; }
-        public string Arch { get; set; }
+        public DroneIntegrity Integrity { get; set; }
+        public DroneArch Arch { get; set; }
         public DateTime LastSeen { get; set; }
+        
+        public enum DroneIntegrity
+        {
+            Medium,
+            High,
+            SYSTEM
+        }
 
+        public enum DroneArch
+        {
+            x86,
+            x64
+        }
+        
         public IEnumerable<DroneModuleResponse> Modules { get; set; }
     }
 }